requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-core==0.2.3
langchain-text-splitters==0.2.0
langsmith==0.1.65
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
orjson==3.10.3
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rsa==4.9
safetensors==0.4.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.2
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
